Added lap time generation
An option to calculate and transfer lap times from eTiming base to LiveRes is made. This is enabled by a check box in the initial dialog. The lap times are added as split times with codes + 100 000. An update to LiveRes web is required for this to show as intended (arrives in next update).